fix: checkpoint integration tests
